@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
+    compileSdkVersion 29
     buildToolsVersion '26.0.2'
 
     defaultConfig {
         applicationId "com.test.banner"
         minSdkVersion 14
-        targetSdkVersion 25
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
@@ -22,9 +22,9 @@ android {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:23+'
-    compile 'com.android.support:recyclerview-v7:23+'
-    compile "com.github.bumptech.glide:glide:3.7.0"
+    compile 'androidx.appcompat:appcompat:1.0.0'
+    compile 'androidx.recyclerview:recyclerview:1.0.0'
+    compile "com.github.bumptech.glide:glide:4.9.0"
     compile 'com.facebook.fresco:fresco:0.12.0'
     compile 'com.zxy.android:recovery:0.0.8'
     compile project(':banner')
