@@ -1,6 +1,6 @@
 package com.test.banner.demo;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.test.banner.App;
