@@ -16,7 +16,7 @@
 
 package com.youth.banner.transformer;
 
-import android.support.v4.view.ViewPager.PageTransformer;
+import androidx.viewpager.widget.ViewPager.PageTransformer;
 import android.view.View;
 
 public abstract class ABaseTransformer implements PageTransformer {
